+++ CHANGELOG.md
@@ -7,6 +7,9 @@ All notable changes to this project will be documented in this file.
 ### Added
 - Version bump to monitor possible HTTP issues releated to ``SetOption128``
 
+### Changed
+- Berry now compiling in ``strict`` mode to catch more bugs
+
 ## [9.5.0.5] 20210815
 ### Added
 - Inital support for Wi-Fi extender (#12784)